Moved tests for endpoint to different classes to simplify test reading
